Some renaming done in knapsack_problem
Mainly I've added prefixes to structure members.
This is to get into the habit of it, so that in large codebases I
can grep for structure member names with less noise.
i.e. searching for 'item_weight' will give me less results than
searching for 'weight', and most of those results will be relevant.