Move the Node type into the users data structures
I imagine that working with this DisjointSet type will not usually be
accessing data elements through the DisjointSet itself.
Instead we will be accessing elements via some other data structure, and
then checking which are of the same set separetely.
Hence I want the functions to find which set you're in to be a trait of
the users data structure, which they can store in whatever collection
they like.