Fundamental problem when multiprocessing
Have found a fundamental problem when multiprocessing with the current
data structure representation.
The only way to get around this is to have an entire data-structure lock
that you obtain when running union(), this breaks the whole point of
making DisjointSet a trait.
I'm hence ending the path of making things multiprocessing here (I
couldn't even find a use for it anyway).