Fix some compilation warnings
Mainly observing that if given a &mut, this can be stored as an
immutable variable, yet still passed to functions that change the
thing that mutable reference is pointing to.

This is analogous to `const char *` vs `char const *`.
The reference may be unable to change but things can be changed through
it, or alternatively the reference may be able to change to point to
something different but things may be unable to change whatever that
reference is pointing to.