Implement .keys() .values() and .remove()
Have gone back to Vec<T> as that has a .remove() method that makes
things easier to implement.

OrderedDict.remove() is currently O(n), but it should be O(1).
Will have to make our own implementation of a Linked List in order to
get the proper complexity profile on the OrderedDict .remove() method.