Tests creating a default OrderedDict is a function
Using a function so I don't have to type out how to create a new Dict
each time I want one to test with.

This will be pretty useful in the future when I split the tests for each
different function and trait implementation into different functions.
(this will be done for more specific error reporting).