Neaten up the simple tests
This commit makes most of the simple tests use the function
assert_iterator_matches(), which means we're testing the size_hint() of
each iterator at the same time as what elements they produce.

In the process of doing this I fixed a problem in OrderedDict::retain()
where the decision of which to retain and which to remove was the wrong
way around.